refactor: adding database.add package
