+++ src/test/java/aicore/suite/UnitSuite.java
@@ -4,6 +4,6 @@ import org.junit.platform.suite.api.SelectPackages;
 import org.junit.platform.suite.api.Suite;
 
 @Suite
-@SelectPackages("aicore.unit")
+@SelectPackages("aicore.unit.database")
 public class UnitSuite {
 }
