+++ src/test/java/aicore/unit/database/add/AddFileUploadDatabaseFormValidationTests.java
@@ -1,4 +1,4 @@
-package aicore.unit.database;
+package aicore.unit.database.add;
 
 import java.io.IOException;
 import java.util.LinkedHashMap;
